requirements.txt
numpy
scipy
matplotlib
scikit-image
scikit-learn
asyncio
aioconsole
flask
requests
ipython
dlib
face_recognition
gtts
pillow
sseclient-py
opencv-python